CH: Deleting moth flocks when all particles die
