+++ Assets/Scripts/MothLifetime.cs.meta
